Added higher order components and refactored the modals and dataDisplay components. Also removed all of the validationRules and put them into one file. Refactored most of the files, still working on the main components and pages.
The reason for the refactor and removal is because I had to overhaul everything anyway due to making everything link to orcid. During the process I found myself copy and pasting a lot and released most of the files were carbon copys with different names, so I created higher order components instead. This means in the future I will only need to change one file for each instead of 12